Update the runtime field in RunDto from Long to Double.

+++ src/main/java/com/rocketden/tester/dto/RunDto.java
@@ -12,5 +12,5 @@ public class RunDto {
     private List<ResultDto> results;
     private Integer numCorrect;
     private Integer numTestCases;
-    private Long runtime;
+    private Double runtime;
 }
